Added Time System
I have added a very simple, yet effective time system. You can easily change the beginning time variables with variables from a previous save game. There is also a format time function that formats the time string for the label. Every 10 minutes in game is equivalent to 2 real time seconds. This can be possibly be changed with the addition of speed controls.