Added Build System
I have added a build system to the game. When you press e, you will be able to select walls to place down. There is also a selection tool that will show if you can build on that tile or not. This selection tool and the build system only work if the build menu is currently open, when you close it, it shuts off. This allows for future implementation of a query building tool.