Build System Fixes
I have fixed some small bugs in the build system. I have created a new autoload script for public varables that will be used across most scripts. I have fixed a problem with being able to place when the indicator shows red, now you can't place a new thing if the tile is occupied, you will need to delete the thing on that tile first. I also added a new wall and removed the vertical wall. I also made it so when you start the game and exit the build menu, you won't be able to place what was last selected.