Updated rotation
Now, when you rotate a tile, it will generate a number from 1 to 4, then based on that it will rotate the tile. However, this presents a problem. On wall tiles, 50 percent of your clicks could potentially rotate the tile and the other 50 percent of your clicks could leave the tile in the same position. With some buildings, there is about a 25% chance that the building will not rotate. Also, there is still a chance that after 100,000,000 clicks, the tile remains in its current position. This can be fixed by recording the rotation of the tile and then setting the rotation to the next subsequent rotation, instead of randomly rotating the tile.