Rotation now works
Instead of generating a random number, it will create an array of data stored in 'tile_data'. This array is structured as [x_flipped, y_flipped, transpose].There is then a string of if else statements to check what the current position of the tile is. It will then flip the tile to the next position. This works with walls, wall corners, and garden tiles.