Updated Build and Rotate Systems
Now, both the build and rotate systems are completely seperate from one another. If you wanted to rotate a tile, you had to press 'e' then 'r'. This made it so you had to enter the build menu, then you were taken to the rotate menu. To simplify both these systems, you can directly go to one of them or switch between them. To enter the build menu, press 'e'. To enter the rotate menu, press 'r'. I have also added labels to the top of the screen to indicate which system you are in (if any).