Added Save System
Now, you can press 'esc' and then hit the save button to save your game. The save game does not use json and instead uses a different way of saving, this is to simplify code. The game automatically loads in your last save. However, it doesn't dave the rotation of the tiles. But, this can easily be achieved.