Fixed Save and Load
When you click save, it now saves if the cell is flipped on the x or y and if the cell is transposed. When the game loads, it uses those saved values and sets them accordingly when it is placing the tiles. This makes sure that whatever cells you rotated stay rotated.